:brain:

+++ tests/test_models/test_engine/test_file_storage.py
@@ -91,13 +91,15 @@ class TestBmFsInstances(unittest.TestCase):
 
     def test_storage_file_exists(self):
         """... checks proper FileStorage instantiation"""
-        os.remove(F)
+        if os.path.isfile(F):
+            os.remove(F)
         self.bm_obj.save()
         self.assertTrue(os.path.isfile(F))
 
     def test_obj_saved_to_file(self):
         """... checks proper FileStorage instantiation"""
-        os.remove(F)
+        if os.path.isfile(F):
+            os.remove(F)
         self.bm_obj.save()
         bm_id = self.bm_obj.id
         actual = 0
@@ -120,7 +122,8 @@ class TestBmFsInstances(unittest.TestCase):
 
     def test_reload(self):
         """... checks proper usage of reload function"""
-        os.remove(F)
+        if os.path.isfile(F):
+            os.remove(F)
         self.bm_obj.save()
         bm_id = self.bm_obj.id
         actual = 0
@@ -134,7 +137,8 @@ class TestBmFsInstances(unittest.TestCase):
 
     def test_save_reload_class(self):
         """... checks proper usage of class attribute in file storage"""
-        os.remove(F)
+        if os.path.isfile(F):
+            os.remove(F)
         self.bm_obj.save()
         bm_id = self.bm_obj.id
         actual = 0
